pyproject.toml
[project]
name = "flowise-document-stores"
version = "2.2.0"
description = "Python toolset for managing Flowise document stores via API - list, refresh, and monitor document stores with async operations"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
keywords = ["flowise", "document-store", "api", "async", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "python-dotenv>=1.0.0",
    "aiohttp>=3.8.0",
]

[project.scripts]
flowise-lister = "flowise_document_lister:main"
flowise-refresher = "flowise_document_refresher:main"
flowise-refresher-auto = "flowise_document_refresher_auto:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"